refactor: move finder + in-editor find/replace into src/finder.rs, src/find.rs
