+++ src/main.rs
@@ -11,7 +11,10 @@ mod divider;
 pub(crate) use divider::{full_island_w, Divider, DIFF_GUTTER_W};
 mod ui;
 pub(crate) use ui::*;
+mod find;
+mod finder;
 mod history;
+pub(crate) use app::{CONTENT_MIN_QUERY, CONTENT_SEARCH_DEBOUNCE, FINDER_RESULT_CAP, SCROLL_CONTEXT_ROWS};
 use kyde_diff as diff;
 mod editor;
 use kyde_config::keymap;
