+++ src/render.rs
@@ -2822,128 +2822,6 @@ impl Kyde {
             .children(blocks)
     }
 
-    /// Find / replace bar shown atop the editor (cmd-f / cmd-r). Live-highlights matches;
-    /// enter / cmd-g cycle, the buttons replace.
-    fn render_find_bar(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let count = self.find_matches.len();
-        let label = if count == 0 {
-            "No results".to_string()
-        } else {
-            format!("{}/{}", self.find_idx + 1, count)
-        };
-        let input_box = |child: gpui::Entity<CodeEditor>| {
-            div()
-                .flex_1()
-                .min_w_0()
-                .h(px(26.0))
-                .px_2()
-                .flex()
-                .items_center()
-                .bg(t.main_bg)
-                .border_1()
-                .border_color(t.divider)
-                .rounded_md()
-                .child(child)
-        };
-        let btn = |glyph: &str, id: &'static str| {
-            div()
-                .id(SharedString::from(id))
-                .size(px(24.0))
-                .flex()
-                .items_center()
-                .justify_center()
-                .rounded_md()
-                .text_color(t.secondary_text)
-                .hover(|s| s.bg(t.bg_light).text_color(t.text))
-                .cursor_pointer()
-                .child(SharedString::from(glyph.to_string()))
-        };
-
-        let find_row = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .child(input_box(self.find_query.clone()))
-            .child(
-                div()
-                    .flex_none()
-                    .text_color(t.line_number)
-                    .text_size(px(theme::get().ui_font_size - 1.0))
-                    .child(label),
-            )
-            .child(btn("‹", "find-prev").on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, w, cx| this.find_prev(&FindPrev, w, cx)),
-            ))
-            .child(btn("›", "find-next").on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, w, cx| this.find_next(&FindNext, w, cx)),
-            ))
-            .child(btn("×", "find-close").on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, w, cx| this.close_find(&CloseFind, w, cx)),
-            ));
-
-        let mut col = div()
-            .flex()
-            .flex_col()
-            .gap_1()
-            .px_2()
-            .py_1p5()
-            .bg(t.panel_bg)
-            .border_b_1()
-            .border_color(t.divider)
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size))
-            .child(find_row);
-
-        if self.find_replace {
-            let replace_row = div()
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_2()
-                .child(input_box(self.replace_query.clone()))
-                .child(
-                    div()
-                        .id("replace-one")
-                        .px_2()
-                        .h(px(24.0))
-                        .flex()
-                        .items_center()
-                        .rounded_md()
-                        .text_color(t.secondary_text)
-                        .hover(|s| s.bg(t.bg_light).text_color(t.text))
-                        .cursor_pointer()
-                        .child("Replace")
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(|this, _e, w, cx| this.replace_one(&ReplaceOne, w, cx)),
-                        ),
-                )
-                .child(
-                    div()
-                        .id("replace-all")
-                        .px_2()
-                        .h(px(24.0))
-                        .flex()
-                        .items_center()
-                        .rounded_md()
-                        .text_color(t.secondary_text)
-                        .hover(|s| s.bg(t.bg_light).text_color(t.text))
-                        .cursor_pointer()
-                        .child("All")
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(|this, _e, w, cx| this.replace_all(&ReplaceAll, w, cx)),
-                        ),
-                );
-            col = col.child(replace_row);
-        }
-        col.into_any_element()
-    }
 
     /// Editor tab strip: one tab per open file, left→right in open order. Click activates,
     /// the `×` closes, right-click opens the tab context menu (close / others / to the right).
@@ -3468,263 +3346,6 @@ impl Kyde {
             .into_any_element()
     }
 
-    fn render_finder(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        // gpui only tracks hover on *identified* elements, so each row needs an `.id()` or the
-        // `.hover(...)` style never applies (the tree rows that hover all carry an id).
-        let row_base = |i: usize, sel: bool| {
-            div()
-                .id(("finder-row", i))
-                .flex()
-                .flex_row()
-                .items_center()
-                // Inset the rounded pill 2px from the panel edges so the hover/selected
-                // background doesn't run into the corners.
-                .mx(px(2.0))
-                .px_3()
-                .py_1()
-                .rounded_md()
-                .cursor_pointer()
-                .text_color(t.text)
-                .when(sel, |d| d.bg(t.selected_bg))
-                // The finder panel itself is `bg_mid`, so hover must be a lighter shade
-                // (`bg_light`) to be visible — `bg_mid` on `bg_mid` shows nothing.
-                .when(!sel, |d| d.hover(|s| s.bg(t.bg_light)))
-        };
-        let rows: Vec<gpui::AnyElement> = match self.finder_mode {
-            FinderMode::Files => self
-                .finder_results
-                .iter()
-                .enumerate()
-                .map(|(i, p)| {
-                    let sel = i == self.finder_selected;
-                    let name: SharedString = p.to_string_lossy().into_owned().into();
-                    let pc = p.clone();
-                    let icon = div()
-                        .w(px(20.0))
-                        .flex_none()
-                        .mr(px(8.0))
-                        .overflow_hidden()
-                        .flex()
-                        .items_center()
-                        .justify_end()
-                        .child(badge_inner(file_badge(p), 0.0));
-                    row_base(i, sel)
-                        .child(icon)
-                        .child(div().min_w_0().truncate().child(name))
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(move |this, _e, window, cx| {
-                                this.open_file(pc.clone(), cx);
-                                this.mode = Mode::Browse;
-                                this.finder_open = false;
-                                window.focus(&this.focus_handle);
-                                cx.notify();
-                            }),
-                        )
-                        .into_any_element()
-                })
-                .collect(),
-            FinderMode::Content => self
-                .content_results
-                .iter()
-                .enumerate()
-                .map(|(i, hit)| {
-                    let sel = i == self.finder_selected;
-                    let path = hit.path.clone();
-                    let line = hit.line;
-                    let loc: SharedString =
-                        format!("{}:{}", hit.path.to_string_lossy(), hit.line).into();
-                    // The matched line, trimmed + capped so a huge minified line can't blow up.
-                    let snippet: SharedString = hit
-                        .text
-                        .trim_start()
-                        .chars()
-                        .take(200)
-                        .collect::<String>()
-                        .into();
-                    let icon = div()
-                        .w(px(20.0))
-                        .flex_none()
-                        .mr(px(8.0))
-                        .overflow_hidden()
-                        .flex()
-                        .items_center()
-                        .justify_end()
-                        .child(badge_inner(file_badge(&path), 0.0));
-                    row_base(i, sel)
-                        .child(icon)
-                        .child(
-                            div()
-                                .flex_1()
-                                .min_w_0()
-                                .truncate()
-                                .font_family(theme::font::FAMILY)
-                                .child(snippet),
-                        )
-                        .child(
-                            div()
-                                .flex_none()
-                                .ml(px(10.0))
-                                .max_w(px(220.0))
-                                .truncate()
-                                .text_color(t.line_number)
-                                .child(loc),
-                        )
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(move |this, _e, window, cx| {
-                                this.finder_open = false;
-                                this.open_file_at_line(path.clone(), line, window, cx);
-                            }),
-                        )
-                        .into_any_element()
-                })
-                .collect(),
-            FinderMode::Actions => self
-                .action_results
-                .iter()
-                .enumerate()
-                .map(|(row, &ai)| {
-                    let (label, kind, key_action) = PALETTE[ai];
-                    let sel = row == self.finder_selected;
-                    // Right-aligned shortcut chip, when this action has a bound key.
-                    let shortcut = (!key_action.is_empty())
-                        .then(|| self.keymap.key_for(key_action))
-                        .flatten()
-                        .map(|k| {
-                            div()
-                                .px_2()
-                                .bg(t.bg_mid)
-                                .rounded_md()
-                                .text_color(t.line_number)
-                                .child(SharedString::from(pretty_key(&k)))
-                        });
-                    row_base(row, sel)
-                        .child(
-                            div()
-                                .flex_1()
-                                .min_w_0()
-                                .truncate()
-                                .child(SharedString::from(label)),
-                        )
-                        .children(shortcut)
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(move |this, _e, window, cx| {
-                                this.run_palette(kind, window, cx)
-                            }),
-                        )
-                        .into_any_element()
-                })
-                .collect(),
-            FinderMode::Scratch => self
-                .action_results
-                .iter()
-                .enumerate()
-                .map(|(row, &li)| {
-                    let (label, ext) = scratch::LANGS[li];
-                    let sel = row == self.finder_selected;
-                    let icon = div()
-                        .w(px(20.0))
-                        .flex_none()
-                        .mr(px(8.0))
-                        .flex()
-                        .items_center()
-                        .justify_end()
-                        .child(badge_inner(
-                            file_badge(std::path::Path::new(&format!("x.{ext}"))),
-                            0.0,
-                        ));
-                    row_base(row, sel)
-                        .child(icon)
-                        .child(SharedString::from(label))
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(move |this, _e, _w, cx| this.create_scratch(ext, cx)),
-                        )
-                        .into_any_element()
-                })
-                .collect(),
-        };
-
-        let modal = div()
-            .key_context("FileFinder")
-            .on_action(cx.listener(Self::finder_up))
-            .on_action(cx.listener(Self::finder_down))
-            .on_action(cx.listener(Self::finder_confirm))
-            .on_action(cx.listener(Self::finder_close))
-            .w(px(640.0))
-            .max_h(px(440.0))
-            .flex()
-            .flex_col()
-            .bg(theme::get().bg_mid)
-            .border_1()
-            .border_color(theme::get().bg_light)
-            .rounded_md()
-            .shadow_lg()
-            .font_family(ui)
-            .text_size(fs)
-            .child(
-                div()
-                    .p_2()
-                    .border_b_1()
-                    .border_color(theme::get().divider)
-                    .child(self.finder_query.clone()),
-            )
-            .when(
-                matches!(self.finder_mode, FinderMode::Content)
-                    && !self.finder_query.read(cx).text().is_empty(),
-                |d| {
-                    let n = self.content_results.len();
-                    let files = self
-                        .content_results
-                        .iter()
-                        .map(|h| &h.path)
-                        .collect::<std::collections::HashSet<_>>()
-                        .len();
-                    let label: SharedString = if n == 0 {
-                        "No matches".into()
-                    } else {
-                        format!(
-                            "{n} match{} in {files} file{}",
-                            if n == 1 { "" } else { "es" },
-                            if files == 1 { "" } else { "s" }
-                        )
-                        .into()
-                    };
-                    d.child(
-                        div()
-                            .px_3()
-                            .py_1()
-                            .border_b_1()
-                            .border_color(t.divider)
-                            .text_color(t.line_number)
-                            .child(label),
-                    )
-                },
-            )
-            .child(
-                div()
-                    .id("finder-results")
-                    .overflow_y_scroll()
-                    .flex()
-                    .flex_col()
-                    // 2px vertical gap between the rounded rows, matching their 2px side inset.
-                    .gap(px(2.0))
-                    .children(rows),
-            )
-            // Clicks inside the panel must not reach the backdrop (which would close the finder
-            // right after a row's handler ran — breaking palette actions that re-open a finder).
-            .on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation());
-
-        overlay(cx, true).child(modal).into_any_element()
-    }
 
     /// The language-plugin manager: every installable pack with its monogram badge,
     /// approximate compiled footprint, install state, and an Install/Uninstall button.
